move remote control bridge impl
